LayoutTests/svg/dom/rgb-color-parser.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/scripted-random.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
    description("This test fuzzes the color parser with semi-random attribute values and dumps the results of any values that parse successfully.");

    var characters = [
        "0",
        "1",
        "2",
        "3",
        "4",
        "5",
        "6",
        "7",
        "8",
        "9",
        " ",
        "\t",
        "(",
        ")",
        "%",
        "+",
        "-",
        ".",
        "e",
        ","
    ];
    
    var stopElement = document.createElementNS("http://www.w3.org/2000/svg", "stop");
    function parseRGBColor(string)
    {
        try {
            stopElement.style.removeProperty("stop-color");
            stopElement.style.stopColor = string;
            var stopColor = stopElement.style.getPropertyCSSValue("stop-color");
            if (!stopColor) {
                debug("Failed to parse: " + string);
                return;
            }
            var rgbColor = stopColor.rgbColor;
            var red = rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER);
            var green = rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER);
            var blue = rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER);
            var rgbColorString = "rgb(" + red + "," + green + "," + blue + ")";
            debug("Parsed as " + rgbColorString + ": " + string);
        } catch(e) {
            debug("Threw exception " + e + ": " + string);
        }
    }

    // Taken from CSS 3 color.
    var svgColors = [
        "aliceblue",
        "antiquewhite",
        "aqua",
        "aquamarine",
        "azure",
        "beige",
        "bisque",
        "black",
        "blanchedalmond",
        "blue",
        "blueviolet",
        "brown",
        "burlywood",
        "cadetblue",
        "chartreuse",
        "chocolate",
        "coral",
        "cornflowerblue",
        "cornsilk",
        "crimson",
        "cyan",
        "darkblue",
        "darkcyan",
        "darkgoldenrod",
        "darkgray",
        "darkgreen",
        "darkgrey",
        "darkkhaki",
        "darkmagenta",
        "darkolivegreen",
        "darkorange",
        "darkorchid",
        "darkred",
        "darksalmon",
        "darkseagreen",
        "darkslateblue",
        "darkslategray",
        "darkslategrey",
        "darkturquoise",
        "darkviolet",
        "deeppink",
        "deepskyblue",
        "dimgray",
        "dimgrey",
        "dodgerblue",
        "firebrick",
        "floralwhite",
        "forestgreen",
        "fuchsia",
        "gainsboro",
        "ghostwhite",
        "gold",
        "goldenrod",
        "gray",
        "green",
        "greenyellow",
        "grey",
        "honeydew",
        "hotpink",
        "indianred",
        "indigo",
        "ivory",
        "khaki",
        "lavender",
        "lavenderblush",
        "lawngreen",
        "lemonchiffon",
        "lightblue",
        "lightcoral",
        "lightcyan",
        "lightgoldenrodyellow",
        "lightgray",
        "lightgreen",
        "lightgrey",
        "lightpink",
        "lightsalmon",
        "lightseagreen",
        "lightskyblue",
        "lightslategray",
        "lightslategrey",
        "lightsteelblue",
        "lightyellow",
        "lime",
        "limegreen",
        "linen",
        "magenta",
        "maroon",
        "mediumaquamarine",
        "mediumblue",
        "mediumorchid",
        "mediumpurple",
        "mediumseagreen",
        "mediumslateblue",
        "mediumspringgreen",
        "mediumturquoise",
        "mediumvioletred",
        "midnightblue",
        "mintcream",
        "mistyrose",
        "moccasin",
        "navajowhite",
        "navy",
        "oldlace",
        "olive",
        "olivedrab",
        "orange",
        "orangered",
        "orchid",
        "palegoldenrod",
        "palegreen",
        "paleturquoise",
        "palevioletred",
        "papayawhip",
        "peachpuff",
        "peru",
        "pink",
        "plum",
        "powderblue",
        "purple",
        // We do not test red.
        "rosybrown",
        "royalblue",
        "saddlebrown",
        "salmon",
        "sandybrown",
        "seagreen",
        "seashell",
        "sienna",
        "silver",
        "skyblue",
        "slateblue",
        "slategray",
        "slategrey",
        "snow",
        "springgreen",
        "steelblue",
        "tan",
        "teal",
        "thistle",
        "tomato",
        "turquoise",
        "violet",
        "wheat",
        "white",
        "whitesmoke",
        "yellow",
        "yellowgreen"
    ];

    function fuzz()
    {
        // Some valid values.
        parseRGBColor("blue");
        parseRGBColor("rgb(0, 255, 0)");
        for (var i = 0; i < svgColors.length; ++i)
            parseRGBColor(svgColors[i]);

        // Some invalid ones.
        parseRGBColor("rgb(100%,100%,0%");
        parseRGBColor("rgba(100%,100%,0%");
        parseRGBColor("rgb(100%,100%,r)");
        parseRGBColor("rgb (100%,100%,r)");
        parseRGBColor("rgb(100%, 1, 1)");
        parseRGBColor("hsl(100%,100%,100%)");
        parseRGBColor("yello");
        parseRGBColor("!#fff");
        parseRGBColor("#sdf");
        parseRGBColor("#");
        parseRGBColor("#f");
        parseRGBColor("#ff");
        parseRGBColor("#fffff");
        parseRGBColor("#fffffff");
        parseRGBColor("green,");
        parseRGBColor("'left'");
        parseRGBColor("'right'");
        parseRGBColor("'center'");
        parseRGBColor("'border'");
        parseRGBColor("'border-");
        parseRGBColor("'content'");
        parseRGBColor("'logical'");
        parseRGBColor("'visual'");

        // Some more valid ones.
        parseRGBColor("rgb(0, 10, 20)");
        parseRGBColor("#fff");
        parseRGBColor("#ffff");
        parseRGBColor("#ffffff");
        parseRGBColor("#ffffffff");
        parseRGBColor("#FFF");
        parseRGBColor("#FFFFFF");
        parseRGBColor(" green");
        parseRGBColor("green ");
        parseRGBColor("green");
        
        // Random assortments of valid characters
        for (var i = 0; i < 250; i++) { //>
            var colorString = "rgb(";
            var count = Math.scriptedRandomInt(20);
            for (var j = 0; j < count; j++) { //>
                colorString += characters[Math.scriptedRandomInt(characters.length)];
            }
            parseRGBColor(colorString);
        }
        
        // Empty-ish colors
        parseRGBColor("");
        parseRGBColor(String.fromCharCode(0));
        parseRGBColor("rgb(" + String.fromCharCode(0) + ")");
        
        // One more valid value.
        parseRGBColor("#0f0");
    }

    fuzz();

</script>
<script src="../../resources/js-test-post.js"></script>
</html>
