LayoutTests/svg/dom/rgb-color-parser-expected.txt

This test fuzzes the color parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Parsed as rgb(0,0,255): blue
Parsed as rgb(0,255,0): rgb(0, 255, 0)
Parsed as rgb(240,248,255): aliceblue
Parsed as rgb(250,235,215): antiquewhite
Parsed as rgb(0,255,255): aqua
Parsed as rgb(127,255,212): aquamarine
Parsed as rgb(240,255,255): azure
Parsed as rgb(245,245,220): beige
Parsed as rgb(255,228,196): bisque
Parsed as rgb(0,0,0): black
Parsed as rgb(255,235,205): blanchedalmond
Parsed as rgb(0,0,255): blue
Parsed as rgb(138,43,226): blueviolet
Parsed as rgb(165,42,42): brown
Parsed as rgb(222,184,135): burlywood
Parsed as rgb(95,158,160): cadetblue
Parsed as rgb(127,255,0): chartreuse
Parsed as rgb(210,105,30): chocolate
Parsed as rgb(255,127,80): coral
Parsed as rgb(100,149,237): cornflowerblue
Parsed as rgb(255,248,220): cornsilk
Parsed as rgb(220,20,60): crimson
Parsed as rgb(0,255,255): cyan
Parsed as rgb(0,0,139): darkblue
Parsed as rgb(0,139,139): darkcyan
Parsed as rgb(184,134,11): darkgoldenrod
Parsed as rgb(169,169,169): darkgray
Parsed as rgb(0,100,0): darkgreen
Parsed as rgb(169,169,169): darkgrey
Parsed as rgb(189,183,107): darkkhaki
Parsed as rgb(139,0,139): darkmagenta
Parsed as rgb(85,107,47): darkolivegreen
Parsed as rgb(255,140,0): darkorange
Parsed as rgb(153,50,204): darkorchid
Parsed as rgb(139,0,0): darkred
Parsed as rgb(233,150,122): darksalmon
Parsed as rgb(143,188,143): darkseagreen
Parsed as rgb(72,61,139): darkslateblue
Parsed as rgb(47,79,79): darkslategray
Parsed as rgb(47,79,79): darkslategrey
Parsed as rgb(0,206,209): darkturquoise
Parsed as rgb(148,0,211): darkviolet
Parsed as rgb(255,20,147): deeppink
Parsed as rgb(0,191,255): deepskyblue
Parsed as rgb(105,105,105): dimgray
Parsed as rgb(105,105,105): dimgrey
Parsed as rgb(30,144,255): dodgerblue
Parsed as rgb(178,34,34): firebrick
Parsed as rgb(255,250,240): floralwhite
Parsed as rgb(34,139,34): forestgreen
Parsed as rgb(255,0,255): fuchsia
Parsed as rgb(220,220,220): gainsboro
Parsed as rgb(248,248,255): ghostwhite
Parsed as rgb(255,215,0): gold
Parsed as rgb(218,165,32): goldenrod
Parsed as rgb(128,128,128): gray
Parsed as rgb(0,128,0): green
Parsed as rgb(173,255,47): greenyellow
Parsed as rgb(128,128,128): grey
Parsed as rgb(240,255,240): honeydew
Parsed as rgb(255,105,180): hotpink
Parsed as rgb(205,92,92): indianred
Parsed as rgb(75,0,130): indigo
Parsed as rgb(255,255,240): ivory
Parsed as rgb(240,230,140): khaki
Parsed as rgb(230,230,250): lavender
Parsed as rgb(255,240,245): lavenderblush
Parsed as rgb(124,252,0): lawngreen
Parsed as rgb(255,250,205): lemonchiffon
Parsed as rgb(173,216,230): lightblue
Parsed as rgb(240,128,128): lightcoral
Parsed as rgb(224,255,255): lightcyan
Parsed as rgb(250,250,210): lightgoldenrodyellow
Parsed as rgb(211,211,211): lightgray
Parsed as rgb(144,238,144): lightgreen
Parsed as rgb(211,211,211): lightgrey
Parsed as rgb(255,182,193): lightpink
Parsed as rgb(255,160,122): lightsalmon
Parsed as rgb(32,178,170): lightseagreen
Parsed as rgb(135,206,250): lightskyblue
Parsed as rgb(119,136,153): lightslategray
Parsed as rgb(119,136,153): lightslategrey
Parsed as rgb(176,196,222): lightsteelblue
Parsed as rgb(255,255,224): lightyellow
Parsed as rgb(0,255,0): lime
Parsed as rgb(50,205,50): limegreen
Parsed as rgb(250,240,230): linen
Parsed as rgb(255,0,255): magenta
Parsed as rgb(128,0,0): maroon
Parsed as rgb(102,205,170): mediumaquamarine
Parsed as rgb(0,0,205): mediumblue
Parsed as rgb(186,85,211): mediumorchid
Parsed as rgb(147,112,219): mediumpurple
Parsed as rgb(60,179,113): mediumseagreen
Parsed as rgb(123,104,238): mediumslateblue
Parsed as rgb(0,250,154): mediumspringgreen
Parsed as rgb(72,209,204): mediumturquoise
Parsed as rgb(199,21,133): mediumvioletred
Parsed as rgb(25,25,112): midnightblue
Parsed as rgb(245,255,250): mintcream
Parsed as rgb(255,228,225): mistyrose
Parsed as rgb(255,228,181): moccasin
Parsed as rgb(255,222,173): navajowhite
Parsed as rgb(0,0,128): navy
Parsed as rgb(253,245,230): oldlace
Parsed as rgb(128,128,0): olive
Parsed as rgb(107,142,35): olivedrab
Parsed as rgb(255,165,0): orange
Parsed as rgb(255,69,0): orangered
Parsed as rgb(218,112,214): orchid
Parsed as rgb(238,232,170): palegoldenrod
Parsed as rgb(152,251,152): palegreen
Parsed as rgb(175,238,238): paleturquoise
Parsed as rgb(219,112,147): palevioletred
Parsed as rgb(255,239,213): papayawhip
Parsed as rgb(255,218,185): peachpuff
Parsed as rgb(205,133,63): peru
Parsed as rgb(255,192,203): pink
Parsed as rgb(221,160,221): plum
Parsed as rgb(176,224,230): powderblue
Parsed as rgb(128,0,128): purple
Parsed as rgb(188,143,143): rosybrown
Parsed as rgb(65,105,225): royalblue
Parsed as rgb(139,69,19): saddlebrown
Parsed as rgb(250,128,114): salmon
Parsed as rgb(244,164,96): sandybrown
Parsed as rgb(46,139,87): seagreen
Parsed as rgb(255,245,238): seashell
Parsed as rgb(160,82,45): sienna
Parsed as rgb(192,192,192): silver
Parsed as rgb(135,206,235): skyblue
Parsed as rgb(106,90,205): slateblue
Parsed as rgb(112,128,144): slategray
Parsed as rgb(112,128,144): slategrey
Parsed as rgb(255,250,250): snow
Parsed as rgb(0,255,127): springgreen
Parsed as rgb(70,130,180): steelblue
Parsed as rgb(210,180,140): tan
Parsed as rgb(0,128,128): teal
Parsed as rgb(216,191,216): thistle
Parsed as rgb(255,99,71): tomato
Parsed as rgb(64,224,208): turquoise
Parsed as rgb(238,130,238): violet
Parsed as rgb(245,222,179): wheat
Parsed as rgb(255,255,255): white
Parsed as rgb(245,245,245): whitesmoke
Parsed as rgb(255,255,0): yellow
Parsed as rgb(154,205,50): yellowgreen
Failed to parse: rgb(100%,100%,0%
Failed to parse: rgba(100%,100%,0%
Failed to parse: rgb(100%,100%,r)
Failed to parse: rgb (100%,100%,r)
Failed to parse: rgb(100%, 1, 1)
Failed to parse: hsl(100%,100%,100%)
Failed to parse: yello
Failed to parse: !#fff
Failed to parse: #sdf
Failed to parse: #
Failed to parse: #f
Failed to parse: #ff
Failed to parse: #fffff
Failed to parse: #fffffff
Failed to parse: green,
Failed to parse: 'left'
Failed to parse: 'right'
Failed to parse: 'center'
Failed to parse: 'border'
Failed to parse: 'border-
Failed to parse: 'content'
Failed to parse: 'logical'
Failed to parse: 'visual'
Parsed as rgb(0,10,20): rgb(0, 10, 20)
Parsed as rgb(255,255,255): #fff
Parsed as rgb(255,255,255): #ffff
Parsed as rgb(255,255,255): #ffffff
Parsed as rgb(255,255,255): #ffffffff
Parsed as rgb(255,255,255): #FFF
Parsed as rgb(255,255,255): #FFFFFF
Parsed as rgb(0,128,0):  green
Parsed as rgb(0,128,0): green 
Parsed as rgb(0,128,0): green
Failed to parse: rgb(.78-2e8e
Failed to parse: rgb(% 
Failed to parse: rgb(7
Failed to parse: rgb(e+
Failed to parse: rgb(
Failed to parse: rgb( -2e9%0.795e0,).	
Failed to parse: rgb(%,4+7).-(%+	8 
Failed to parse: rgb(74).9 +(.% 0	635
Failed to parse: rgb()9,- 8	7%4%9%%.
Failed to parse: rgb(49),,4-30378 052
Failed to parse: rgb(%7863	
Failed to parse: rgb(e0
Failed to parse: rgb(%72+e%e,49 7 
Failed to parse: rgb(	55,.
Failed to parse: rgb(263,%	1-74)e
Failed to parse: rgb( 64110, (1 
Failed to parse: rgb( 59%3   725)
Failed to parse: rgb(6e)-2- 180
Failed to parse: rgb(,	.579e(( 
Failed to parse: rgb()(
Failed to parse: rgb(74	e)e6e)-29e7
Failed to parse: rgb(221%  (e(77
Failed to parse: rgb(--4	e1%- )	8378 
Failed to parse: rgb(e4(01.e	
Failed to parse: rgb(
Failed to parse: rgb()6((9((.9 3-6438
Failed to parse: rgb(),15%-+)(,).6
Failed to parse: rgb(3	(-085932
Failed to parse: rgb((
Failed to parse: rgb()
Failed to parse: rgb(e0944 6)
Failed to parse: rgb( 
Failed to parse: rgb(e5
Failed to parse: rgb(3%+)2%3%
Failed to parse: rgb(
Failed to parse: rgb(6e5%.3	e+ e1.69e16
Failed to parse: rgb(1		5.56,+8 5e%,3
Failed to parse: rgb(	)ee.%)
Failed to parse: rgb(-%0%-
Failed to parse: rgb(231293 16%
Failed to parse: rgb(+-2+37)888580+9	(
Failed to parse: rgb(18	805
Failed to parse: rgb(0
Failed to parse: rgb(63-.693+
Failed to parse: rgb(80,2-(.4		,
Failed to parse: rgb(5%(59%2	5
Failed to parse: rgb(5e3%,9
Failed to parse: rgb(%7e-++59e
Failed to parse: rgb(963 (8)%73)	ee(-
Failed to parse: rgb(71+1+()3382 3
Failed to parse: rgb(.27+
Failed to parse: rgb(
Failed to parse: rgb((
Failed to parse: rgb(88
Failed to parse: rgb(9.43,(6.87 -
Failed to parse: rgb(e)e. 5997
Failed to parse: rgb(5 9-557
Failed to parse: rgb((65-%-0,e217
Failed to parse: rgb( 2
Failed to parse: rgb((e
Failed to parse: rgb(	8(+((e%52, .73
Failed to parse: rgb(,0(e 
Failed to parse: rgb()59)62159544,8,e3e
Failed to parse: rgb(0857-.5
Failed to parse: rgb(6,71(5	.-%.
Failed to parse: rgb(2	5e
Failed to parse: rgb(1	2-7	+-e90
Failed to parse: rgb(9 )-7
Failed to parse: rgb(8	759 6,
Failed to parse: rgb(2+65e 59%0(8)5
Failed to parse: rgb( 0712
Failed to parse: rgb(	97,	33)-6,+%0)1e(
Failed to parse: rgb( .3.6 	,-57%7 e+6
Failed to parse: rgb(1)7 +%+8 	)-	 -.63e
Failed to parse: rgb(6786).(8912%
Failed to parse: rgb(-	%8-,)886+- 9e))e
Failed to parse: rgb( ,	5732 
Failed to parse: rgb(18)(
Failed to parse: rgb(3,+13%(711%9%	%0 -
Failed to parse: rgb(6e
Failed to parse: rgb(-e(69
Failed to parse: rgb(,.19%e-2-(0)+9-2
Failed to parse: rgb(1%,))(810-5  %37
Failed to parse: rgb(35.-(%
Failed to parse: rgb(362020+-8347081%78
Failed to parse: rgb(-(
Failed to parse: rgb(%) 54+(006e3141%
Failed to parse: rgb(	-e 1(9	 0-7
Failed to parse: rgb(61-2	 
Failed to parse: rgb(7075)5+	67-(+)7-7-
Failed to parse: rgb(	..
Failed to parse: rgb(88-+3%21
Failed to parse: rgb(3%)003+4
Failed to parse: rgb(682).e19)0-4642
Failed to parse: rgb(6.7
Failed to parse: rgb( 9	1 %6(,%0)e20%e-9
Failed to parse: rgb(	(
Failed to parse: rgb(96,	2e0.1%2)5-
Failed to parse: rgb(.8)7+3 
Failed to parse: rgb(
Failed to parse: rgb(8.78
Failed to parse: rgb(2e8e2% 172e+0. -
Failed to parse: rgb(e9
Failed to parse: rgb(0.795e0,).	%%,
Failed to parse: rgb(+7).
Failed to parse: rgb((%+	8 -74).9 +(.
Failed to parse: rgb( 0	635+)9,- 8	
Failed to parse: rgb(%4%9%%.
Failed to parse: rgb(49),,4-30378 052
Failed to parse: rgb(%7863	
Failed to parse: rgb(e0
Failed to parse: rgb(%72+e%e,49 7 
Failed to parse: rgb(	55,.
Failed to parse: rgb(263,%	1-74)e
Failed to parse: rgb( 64110, (1 
Failed to parse: rgb( 59%3   725)
Failed to parse: rgb(6e)-2- 180
Failed to parse: rgb(,	.579e(( 
Failed to parse: rgb()(
Failed to parse: rgb(74	e)e6e)-29e7
Failed to parse: rgb(221%  (e(77
Failed to parse: rgb(--4	e1%- )	8378 
Failed to parse: rgb(e4(01.e	
Failed to parse: rgb(
Failed to parse: rgb()6((9((.9 3-6438
Failed to parse: rgb(),15%-+)(,).6
Failed to parse: rgb(3	(-085932
Failed to parse: rgb((
Failed to parse: rgb()
Failed to parse: rgb(e0944 6)
Failed to parse: rgb( 
Failed to parse: rgb(e5
Failed to parse: rgb(3%+)2%3%
Failed to parse: rgb(
Failed to parse: rgb(6e5%.3	e+ e1.69e16
Failed to parse: rgb(1		5.56,+8 5e%,3
Failed to parse: rgb(	)ee.%)
Failed to parse: rgb(-%0%-
Failed to parse: rgb(231293 16%
Failed to parse: rgb(+-2+37)888580+9	(
Failed to parse: rgb(18	805
Failed to parse: rgb(0
Failed to parse: rgb(63-.693+
Failed to parse: rgb(80,2-(.4		,
Failed to parse: rgb(5%(59%2	5
Failed to parse: rgb(5e3%,9
Failed to parse: rgb(%7e-++59e
Failed to parse: rgb(963 (8)%73)	ee(-
Failed to parse: rgb(71+1+()3382 3
Failed to parse: rgb(.27+
Failed to parse: rgb(
Failed to parse: rgb((
Failed to parse: rgb(88
Failed to parse: rgb(9.43,(6.87 -
Failed to parse: rgb(e)e. 5997
Failed to parse: rgb(5 9-557
Failed to parse: rgb((65-%-0,e217
Failed to parse: rgb( 2
Failed to parse: rgb((e
Failed to parse: rgb(	8(+((e%52, .73
Failed to parse: rgb(,0(e 
Failed to parse: rgb()59)62159544,8,e3e
Failed to parse: rgb(0857-.5
Failed to parse: rgb(6,71(5	.-%.
Failed to parse: rgb(2	5e
Failed to parse: rgb(1	2-7	+-e90
Failed to parse: rgb(9 )-7
Failed to parse: rgb(8	759 6,
Failed to parse: rgb(2+65e 59%0(8)5
Failed to parse: rgb( 0712
Failed to parse: rgb(	97,	33)-6,+%0)1e(
Failed to parse: rgb( .3.6 	,-57%7 e+6
Failed to parse: rgb(1)7 +%+8 	)-	 -.63e
Failed to parse: rgb(6786).(8912%
Failed to parse: rgb(-	%8-,)886+- 9e))e
Failed to parse: rgb( ,	5732 
Failed to parse: rgb(18)(
Failed to parse: rgb(3,+13%(711%9%	%0 -
Failed to parse: rgb(6e
Failed to parse: rgb(-e(69
Failed to parse: rgb(,.19%e-2-(0)+9-2
Failed to parse: rgb(1%,))(810-5  %37
Failed to parse: rgb(35.-(%
Failed to parse: rgb(362020+-8347081%78
Failed to parse: rgb(-(
Failed to parse: rgb(%) 54+(006e3141%
Failed to parse: rgb(	-e 1(9	 0-7
Failed to parse: rgb(61-2	 
Failed to parse: rgb(7075)5+	67-(+)7-7-
Failed to parse: rgb(	..
Failed to parse: rgb(88-+3%21
Failed to parse: rgb(3%)003+4
Failed to parse: rgb(682).e19)0-4642
Failed to parse: rgb(6.7
Failed to parse: rgb( 9	1 %6(,%0)e20%e-9
Failed to parse: rgb(	(
Failed to parse: rgb(96,	2e0.1%2)5-
Failed to parse: rgb(.8)7+3 
Failed to parse: rgb(
Failed to parse: rgb(8.78
Failed to parse: rgb(2e8e2% 172e+0. -
Failed to parse: rgb(e9
Failed to parse: rgb(0.795e0,).	%%,
Failed to parse: rgb(+7).
Failed to parse: rgb((%+	8 -74).9 +(.
Failed to parse: rgb( 0	635+)9,- 8	
Failed to parse: rgb(%4%9%%.
Failed to parse: rgb(49),,4-30378 052
Failed to parse: rgb(%7863	
Failed to parse: rgb(e0
Failed to parse: rgb(%72+e%e,49 7 
Failed to parse: rgb(	55,.
Failed to parse: rgb(263,%	1-74)e
Failed to parse: rgb( 64110, (1 
Failed to parse: rgb( 59%3   725)
Failed to parse: rgb(6e)-2- 180
Failed to parse: rgb(,	.579e(( 
Failed to parse: rgb()(
Failed to parse: rgb(74	e)e6e)-29e7
Failed to parse: rgb(221%  (e(77
Failed to parse: rgb(--4	e1%- )	8378 
Failed to parse: rgb(e4(01.e	
Failed to parse: rgb(
Failed to parse: rgb()6((9((.9 3-6438
Failed to parse: rgb(),15%-+)(,).6
Failed to parse: rgb(3	(-085932
Failed to parse: rgb((
Failed to parse: rgb()
Failed to parse: rgb(e0944 6)
Failed to parse: rgb( 
Failed to parse: rgb(e5
Failed to parse: rgb(3%+)2%3%
Failed to parse: rgb(
Failed to parse: rgb(6e5%.3	e+ e1.69e16
Failed to parse: rgb(1		5.56,+8 5e%,3
Failed to parse: rgb(	)ee.%)
Failed to parse: rgb(-%0%-
Failed to parse: rgb(231293 16%
Failed to parse: rgb(+-2+37)888580+9	(
Failed to parse: rgb(18	805
Failed to parse: rgb(0
Failed to parse: rgb(63-.693+
Failed to parse: rgb(80,2-(.4		,
Failed to parse: rgb(5%(59%2	5
Failed to parse: rgb(5e3%,9
Failed to parse: rgb(%7e-++59e
Failed to parse: rgb(963 (8)%73)	ee(-
Failed to parse: rgb(71+1+()3382 3
Failed to parse: rgb(.27+
Failed to parse: rgb(
Failed to parse: 
Failed to parse: 
Failed to parse: rgb()
Parsed as rgb(0,255,0): #0f0
PASS successfullyParsed is true

TEST COMPLETE
